reorganized code (no-verify)
