+++ logprep/processor/calculator/ast/exceptions.py
@@ -1,0 +1,22 @@
+from logprep.abc.exceptions import LogprepException
+
+
+class CalculatorError(LogprepException): ...
+
+
+class InvalidSyntaxError(CalculatorError): ...
+
+
+class ParsingError(CalculatorError): ...
+
+
+class MissingValueError(CalculatorError): ...
+
+
+class UnknownFunctionError(CalculatorError): ...
+
+
+class EvaluationError(CalculatorError): ...
+
+
+class DivisionByZeroError(CalculatorError): ...
